Implement support for singletouch mode
As opposed to the original driver from intel, this is using all 16 touch
data buffers, which works just fine. The ME detects that 16 buffers are
available, and fills them with data. This allows us to use a common
handler for both, multitouch and singletouch data.

Since the doorbell doesn't get incremented in singletouch mode, we
capture the event that is used instead and increment the doorbell
manually, to signal to the HID thread that new data was received.

Note: This currently lacks an easy way to switch between modes. You have
to change the like `ipts->mode = IPTS_SENSOR_MODE_MULTITOUCH` in
control.c and recompile the module. A parameter based solution will be
added later.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>