Don't store the active stylus tool globally
HID is a stateful protocol so we don't have to store the active tool. We
can just send values for both tools at the same time. These values are
calculated on the fly.

The other method had the advantage that the removal of the old tool was
done in a seperate frame. But it is *very* hard to switch between pen
and eraser without having at least one frame without any proximity at
all. Even if it is possible, userspace should hopefully be able to pick
it up correctly.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>