Begin moving IPTS processing into userspace
To implement multitouch processing properly, it needs to be done in
userspace, because it heavily relies on floating points. The stylus
needs floating points as well, and even though we solved that problem by
using fixed point, using actual floating points would be way better.

If the data is processed in userspace, a large amount of the code in the
driver becomes pointless. The entire data processing loop can be
removed, only the event loop is needed. That can even be
more slimmed down by removing support for singletouch mode, which is not
needed if a proper processing for multitouch can be implemented.

Additionally, the event loop has been refactored even more: It's not a
thread anymore, but it gets registered using a callback to the MEI
subsystem, so the code only runs when a message has been received from
the ME. The error handling was improved and streamlined as well. No more
random return codes! MEI Bus errors are handled as well and cause the
driver to cease doing anything until you reload it.

Last but not least, this adds the actual user space interface to the now
streamlined driver. It is a modified version of the one that @qzed
already used in the feature/fwd-to-userspc branch. Instead of pushing
incoming touch data into a ring buffer and overwriting the data that's
already there, the feedback sending is now tied to someone having read
the buffer from userspace. When a client has read the entire touch data
buffer, feedback is sent for it to tell the ME to refill it. The next
call to read() will return data from the next touch data buffer and so
on.

By doing this we reduce complexicity (no need for a thread that
continously checks the doorbell value, no ring buffer), and make sure
that userspace - when properly implemented - will never miss a frame. I
am not sure how this will impact latency though, some basic tests show
it to be pretty fast though (the test tool + dmesg -w maxed out the
CPU).

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>