Silence weird error from sending feedback
A small UAPI test program written in Go, that just reads the ipts data,
managed to cause an interesting bug: The mei_cldev_send function would
error and return -4, which seems to be the code for "NOT_READY".

The same program written in C didn't trigger the error...

Since the feedback still worked, I think it is ok to silence this one.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>