Don't forward hardware data directly to userspace
Exporting the ipts_device_info struct to userspace doesn't seem like
the best thing to do, since we have no control over its definition, it
is dictated by hardware.

Additionally, there might be values that userspace doesn't care about,
like the size of a feedback buffer.

Instead use a wrapper struct that contains only the values that are
relevant for userspace. That way we can keep the ioctl stable, even if
it ever turns out to be neccessary to change the ipts_device_info
struct.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>