Update README
This aligns the README of this driver with the README of the userspace
iptsd daemon.

Since the driver doesn't do a lot anymore, it doesn't really make sense
to keep the list of what works and what doesn't in this README. A new
list was added to the iptsd README instead.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>